refactor: Move fixtures to conftest.py
