+++ products/batch_exports/backend/tests/temporal/destinations/redshift/conftest.py
@@ -6,7 +6,9 @@ import pytest
 from django.conf import settings
 
 import psycopg
+import aioboto3
 
+from products.batch_exports.backend.temporal.destinations.redshift_batch_export import Credentials
 from products.batch_exports.backend.tests.temporal.destinations.redshift.utils import MISSING_REQUIRED_ENV_VARS
 
 
@@ -75,3 +77,72 @@ def properties_data_type(request) -> str:
         return request.param
     except AttributeError:
         return "varchar"
+
+
+@pytest.fixture
+def bucket_name() -> str | None:
+    """Name for a test S3 bucket."""
+    test_bucket = os.getenv("S3_TEST_BUCKET")
+
+    if not test_bucket:
+        return None
+
+    return test_bucket
+
+
+@pytest.fixture
+def bucket_region() -> str | None:
+    """Region for a test S3 bucket."""
+    bucket_region = os.getenv("AWS_REGION")
+
+    if not bucket_region:
+        return None
+
+    return bucket_region
+
+
+@pytest.fixture
+def aws_credentials() -> Credentials | None:
+    """AWS credentials to test Redshift copy activity with an S3 bucket."""
+    aws_access_key_id, aws_secret_access_key = os.getenv("AWS_ACCESS_KEY_ID"), os.getenv("AWS_SECRET_ACCESS_KEY")
+
+    if not aws_access_key_id or not aws_secret_access_key:
+        return None
+
+    return Credentials(aws_access_key_id, aws_secret_access_key)
+
+
+@pytest.fixture
+def key_prefix(ateam) -> str:
+    return f"/test-copy-redshift-batch-export_{ateam.pk}"
+
+
+@pytest.fixture
+async def s3_client(credentials, bucket_name):
+    """Manage an S3 client to interact with an S3 bucket."""
+    if not credentials or not bucket_name:
+        return
+
+    async with aioboto3.Session().client("s3") as s3_client:
+        yield s3_client
+
+
+async def delete_all_from_s3(s3_client, bucket_name: str, key_prefix: str):
+    """Delete all objects in bucket_name under key_prefix."""
+    response = await s3_client.list_objects_v2(Bucket=bucket_name, Prefix=key_prefix)
+
+    if "Contents" in response:
+        for obj in response["Contents"]:
+            if "Key" in obj:
+                await s3_client.delete_object(Bucket=bucket_name, Key=obj["Key"])
+
+
+@pytest.fixture(autouse=True)
+async def clean_up_s3_bucket(s3_client, bucket_name, key_prefix):
+    """Clean-up S3 bucket used in Redshift copy activity."""
+    yield
+
+    if not s3_client:
+        return
+
+    await delete_all_from_s3(s3_client, bucket_name, key_prefix)
